@@ -1,41 +1,23 @@
 {
   "name": "my-app",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
-    "@radix-ui/react-menubar": "^1.0.4",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@reduxjs/toolkit": "^2.2.5",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
-    "lucide-react": "^0.394.0",
-    "next": "14.2.4",
-    "react": "^18",
-    "react-dom": "^18",
-    "react-redux": "^9.1.2",
-    "tailwind-merge": "^2.3.0",
-    "tailwindcss-animate": "^1.0.7"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "test": "vitest --coverage"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.4",
-    "postcss": "^8.4.38",
-    "tailwindcss": "^3.4.4",
-    "typescript": "^5"
+    "@vitest/coverage-v8": "^1.6.0",
+    "typescript": "^5.2.2",
+    "vite": "^5.2.0",
+    "vitest": "^1.6.0"
   },
-  "description": "This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).",
-  "main": "index.js",
-  "author": "",
-  "license": "ISC"
+  "dependencies": {
+    "@reduxjs/toolkit": "^2.2.5",
+    "autoprefixer": "^10.4.19",
+    "tailwindcss": "^3.4.4"
+  }
 }
